requirements.txt
# requirements.txt
nats-py==2.6.0
python-dotenv==1.0.0
opencv-python
numpy
ultralytics
asyncio

